Tests/LibWeb/Text/expected/css/animation-inherited-style-update-descendant-slot-scan-count.txt

deep child color: rgb(128, 0, 128)
descendant slot scans stayed linear: true
